Restructure to prepare for a tree-walking interpreter
