+++ src/ast.rs
@@ -1,10 +1,5 @@
 use ess::Sexp;
 
-use std::borrow::{Cow, ToOwned};
-use std::collections::HashMap;
-use std::rc::Rc;
-use std::cell::RefCell;
-
 #[derive(Debug)]
 pub enum AstError<'a> {
     ArgList(Sexp<'a>),
@@ -14,13 +9,6 @@ pub enum AstError<'a> {
     InvalidInCall(Sexp<'a>),
     StructName(Sexp<'a>),
     StructMember(Sexp<'a>),
-    NameLookup(Cow<'a, str>),
-}
-
-fn extend_cow<'a, T: ?Sized>(cow: &Cow<'a, T>) -> Cow<'static, T>
-    where T: ToOwned
-{
-    Cow::Owned(cow.clone().into_owned())
 }
 
 impl<'a> AstError<'a> {
@@ -34,19 +22,15 @@ impl<'a> AstError<'a> {
             InvalidInCall(ref s) => InvalidInCall(s.to_owned()),
             StructName(ref s) => StructName(s.to_owned()),
             StructMember(ref s) => StructMember(s.to_owned()),
-            NameLookup(ref s) => NameLookup(extend_cow(s)),
         }
     }
 }
-
-type NameId = u64;
 
 #[derive(Debug, Clone)]
 pub enum Ast<'a> {
     Call(Box<Ast<'a>>, Vec<Ast<'a>>),
     If(Box<Ast<'a>>, Box<Ast<'a>>, Box<Ast<'a>>),
     Let(Box<Ast<'a>>, Box<Ast<'a>>),
-    Set(Box<Ast<'a>>, Box<Ast<'a>>),
     Begin(Vec<Ast<'a>>),
     Loop(Vec<Ast<'a>>),
     Return(Option<Box<Ast<'a>>>),
@@ -55,38 +39,8 @@ pub enum Ast<'a> {
     Defn(&'a str, Vec<&'a str>, Vec<Ast<'a>>),
     Struct(&'a str, Vec<&'a str>),
     Var(&'a str),
-    Name(NameId),
     Literal(&'a Sexp<'a>),
     Nil,
-}
-
-#[derive(Debug)]
-pub struct Env<'a> {
-    parent: Option<&'a Env<'a>>,
-    names: HashMap<&'a str, NameId>,
-    counter: Rc<RefCell<NameId>>,
-}
-
-impl<'a> Env<'a> {
-    pub fn new() -> Self {
-        Env {
-            parent: None,
-            names: HashMap::new(),
-            counter: Rc::new(RefCell::new(0)),
-        }
-    }
-
-    pub fn with_parent<'b>(parent: &'a Env<'b>) -> Env<'a> {
-        Env {
-            parent: Some(parent),
-            names: HashMap::new(),
-            counter: parent.counter.clone(),
-        }
-    }
-
-    pub fn lookup<'b>(&self, name: &'b str) -> Result<NameId, AstError<'b>> {
-        Err(AstError::NameLookup(name.into()))
-    }
 }
 
 fn defn_get_args<'a>(arg_list: &'a Sexp<'a>)
@@ -127,12 +81,6 @@ pub fn make_ast<'a>(sexp: &'a Sexp<'a>) -> Result<Ast<'a>, AstError<'a>> {
                     let var = make_ast(&xs[1])?;
                     let val = make_ast(&xs[2])?;
                     Ok(Ast::Let(Box::new(var), Box::new(val)))
-                }
-                Sexp::Sym(ref s, ..) if s == "set!" => {
-                    // TODO: Not too many args
-                    let var = make_ast(&xs[1])?;
-                    let val = make_ast(&xs[2])?;
-                    Ok(Ast::Set(Box::new(var), Box::new(val)))
                 }
                 Sexp::Sym(ref s, ..) if s == "begin" => {
                     let body = xs[1..].iter().map(make_ast)
@@ -213,71 +161,3 @@ pub fn make_ast<'a>(sexp: &'a Sexp<'a>) -> Result<Ast<'a>, AstError<'a>> {
         ref s @ Sexp::Float(..) => Ok(Ast::Literal(s)),
     }
 }
-
-/*
-pub fn resolve_names<'a, 'b>(env: &'a Env<'a>, ast: &Ast<'b>) -> Result<Ast<'b>, AstError<'b>> {
-    match *ast {
-        Ast::Var(ref name) => env.lookup(&name).map(Ast::Name),
-        Ast::Defn(_, _, _) => unimplemented!(),
-        Ast::Struct(ref name, ref members) => {
-            unimplemented!()
-        }
-        Ast::Call(ref func, ref args) => {
-            let mut new_args = Vec::with_capacity(args.len());
-            for arg in args {
-                new_args.push(resolve_names(&env, arg)?);
-            }
-            Ok(Ast::Call(Box::new(resolve_names(env, func)?), new_args))
-        }
-        Ast::If(ref cond, ref if_true, ref if_false) => {
-            let cond = resolve_names(env, cond)?;
-            let if_true = resolve_names(env, if_true)?;
-            let if_false = resolve_names(env, if_false)?;
-            Ok(Ast::If(Box::new(cond), Box::new(if_true), Box::new(if_false)))
-        }
-        Ast::Let(ref var, ref val) => {
-            let var = resolve_names(env, var)?;
-            let val = resolve_names(env, val)?;
-            // TODO: Add var to the env if appropriate
-            Ok(Ast::Let(Box::new(var), Box::new(val)))
-        }
-        Ast::Set(ref var, ref val) => {
-            let var = resolve_names(env, var)?;
-            let val = resolve_names(env, val)?;
-            Ok(Ast::Set(Box::new(var), Box::new(val)))
-        }
-        Ast::Begin(ref body) => {
-            let mut env = Env::with_parent(env);
-            let mut new_body = Vec::with_capacity(body.len());
-            for item in body {
-                new_body.push(resolve_names(&env, item)?);
-            }
-            Ok(Ast::Begin(new_body))
-        }
-        Ast::Loop(ref body) => {
-            let mut env = Env::with_parent(env);
-            let mut new_body = Vec::with_capacity(body.len());
-            for item in body {
-                new_body.push(resolve_names(&env, item)?);
-            }
-            Ok(Ast::Loop(new_body))
-        }
-        Ast::Return(ref s) => {
-            match *s {
-                Some(ref x) => Ok(Ast::Return(Some(Box::new(resolve_names(env, x)?)))),
-                None => Ok(Ast::Return(None)),
-            }
-        }
-        Ast::Break(ref s) => {
-            match *s {
-                Some(ref x) => Ok(Ast::Break(Some(Box::new(resolve_names(env, x)?)))),
-                None => Ok(Ast::Break(None)),
-            }
-        }
-        Ast::Continue => Ok(Ast::Continue),
-        ref lit@Ast::Literal(_) => Ok(lit.clone()),
-        Ast::Name(name) => Ok(Ast::Name(name)),
-        Ast::Nil => Ok(Ast::Nil),
-    }
-}
-*/
