+++ src/eval.rs
@@ -1,0 +1,2 @@
+#[derive(Debug, Clone)]
+pub struct Var(());
